uint32 for clear depth

+++ cocos/core/renderer/core/memory-pools.ts
@@ -773,7 +773,7 @@ const cameraViewDataType: BufferDataTypeManifest<typeof CameraView> = {
     [CameraView.EXPOSURE]: BufferDataType.FLOAT32,
     [CameraView.CLEAR_FLAG]: BufferDataType.UINT32,
     [CameraView.CLEAR_DEPTH]: BufferDataType.FLOAT32,
-    [CameraView.CLEAR_STENCIL]: BufferDataType.FLOAT32,
+    [CameraView.CLEAR_STENCIL]: BufferDataType.UINT32,
     [CameraView.NODE]: BufferDataType.UINT32,
     [CameraView.SCENE]: BufferDataType.UINT32,
     [CameraView.FRUSTUM]: BufferDataType.UINT32,
